views divided to pages and endpoints
