+++ mast/__init__.py
@@ -22,4 +22,4 @@ login_manager = LoginManager(app)
 login_manager.login_view = 'login'
 login_manager.login_message_category = 'info'
 
-from mast import views
+from mast import views_pages, views_endpoints
