+++ mast/views_pages.py
@@ -1,13 +1,11 @@
-import json
 import os
 import datetime
 import mast
-from flask import redirect, request, render_template, url_for, jsonify, Blueprint
+from flask import redirect, request, render_template, url_for
 from flask_login import login_user, current_user, logout_user, login_required
 from werkzeug.utils import secure_filename
 from mast.forms import LoginForm, RegisterForm, UpdateProfileForm, ChangePasswordForm, AddActivityForm
 from mast.models import User, Competition, Sex, Age, Activity, ActivityType
-from mast.json_encoder import MastEncoder
 from mast import bcr, queries, app, session
 from mast.tools.sis_authentication import authenticate_via_sis
 from mast.processor import GPXProcessor
@@ -79,8 +77,6 @@ def logout():
 def home():
     session_data = mast.session.Session()
     db_query = mast.queries.Queries()
-    last_activities = db_query.get_user_last_activities(current_user.id, 10)
-    last_activities = [] if not last_activities else last_activities
     add_activity_form = AddActivityForm()
     if add_activity_form.validate_on_submit():
         filename = secure_filename(add_activity_form.file.data.filename)
@@ -130,26 +126,6 @@ def home():
                            season=db_query.SEASON, session_data=session_data)
 
 
-@app.route('/get_personal_stats')
-@login_required
-def get_personal_stats():
-    db_query = mast.queries.Queries()
-    data = db_query.get_total_distances_by_user_in_last_days(user_id=current_user.id, days=7)
-    labels = [key for key, val in data.items()]
-    data = [val for key, val in data.items()]
-    return jsonify({'payload': json.dumps({'data': data, 'labels': labels})})
-
-
-@app.route('/get_personal_activities')
-@login_required
-def get_personal_activities():
-    db_query = mast.queries.Queries()
-    offset = request.args.get('offset')
-    limit = request.args.get('limit')
-    data = db_query.get_user_last_activities(user_id=current_user.id, number=limit, offset=offset)
-    return json.dumps({'total': data[0], 'rows': data[1]}, cls=MastEncoder)
-
-
 @app.route('/matfyz_challenges')
 @login_required
 def matfyz_challenges():
@@ -166,18 +142,6 @@ def matfyz_challenges():
     return render_template("matfyz_challenges.html", title='Matfyz Challenges',
                            checkpoints=checkpoints_enriched, current_checkpoint=current_checkpoint,
                            session_data=session_data)
-
-
-@app.route('/get_global_contest')
-@login_required
-def get_global_contest():
-    db_query = mast.queries.Queries()
-    labels = ["Where we gonna make it by bike.",
-              "Where we gonna make it on foot."]
-    data = [round(db_query.get_global_total_distance_on_bike(), 1),
-            round(db_query.get_global_total_distance_on_foot(), 1)]
-    checkpoints = db_query.get_challenge_parts_to_display()
-    return jsonify({'payload': json.dumps({'data': data, 'labels': labels, 'checkpoints': checkpoints})})
 
 
 @app.route('/running_5_km')
